+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
+## [0.2.0] - 2026-02-25
+
+### Added
+- Helper functions for altering cohort objects in consistent ways (enforcing first event, prior observation period, etc.)
+- Default behaviour of circe models is now better with empty lists (consistent with Java implementation)
+- Additional validation checks for cohort expression objects
+
 ## [0.1.0] - 2026-01-23
 
 ### Added